TST-793 — Add Manage Subscriptions tab to Account Area
This update creates a new custom layout file allowing us to add an unlimited number of new tabs to the Customer Account Area. The page needs to be created in the BigCommerce Admin using the template "Account". The new tab should be manually added to the navigation.html file. Only logged-in customers will be able to see the contents of these pages. Any non-logged access will cause a redirection to the home page.