TST-951 — Fix subscription modal not closing on PDP page
This is a bug with the Modal Factory where the modal options are set globally when you set just one modal. Because of this, the close on background and close on esc options need to be always set BEFORE the modal is open, rather than setting them on the modal creation.